refactor: PR notices fixed
