+++ doc/refs.bib
@@ -33,11 +33,9 @@
   ClinchGaramvölgyiEtAl2024,
   author = {Katie Clinch and Dániel Garamvölgyi and John Haslegrave and Tony Huynh and Jan Legerský and Anthony Nixon},
   title = {Stable cuts, {{NAC}}-colourings and flexible realisations of graphs},
+  archivePrefix = {arXiv},
   year = {2024},
   eprint = {2412.16018},
-  archivePrefix = {arXiv},
-  primaryClass = {math.CO},
-  url = {https://arxiv.org/abs/2412.16018},
 }
 
 @article{
