+++ pyrigi/__init__.py
@@ -5,15 +5,6 @@ This is a package for the rigidity and flexibility of graphs
 and bar-and-joint frameworks.
 """
 
-# Do not export internal functions yet!
-# Internal functions are not exported until the graph's monolit design is decomposed.
-# To use pyrigi.Graph class safely in external functions, typing has to be used.
-# As these functions are referenced also in the Graph clss itself, a dependency
-# cycles is formed. Therefore, (probably) the best approach is to use
-# nx.Graph in external functions for now and call Graph's methods
-# in unsafe way till the monolit is decomposed and we can switch
-# directly to calling new functions (previous methods) directly.
-
 from pyrigi.graph import Graph
 from pyrigi.framework import Framework
 from pyrigi.graph_drawer import GraphDrawer
